+++ server/templates/base.html
@@ -108,8 +108,8 @@
 >
 </div>
 
-{% include 'auxiliary/labels.html' %}
-{% include 'auxiliary/routes.html' %}
+{% include 'metadata/labels.html' %}
+{% include 'metadata/routes.html' %}
 
 <script src={{url_for('static', filename='base.js', t=config['VERSION'])}} async></script>
   <div id="main">
