+++ static/js/apps/base/components/footer.tsx
