+++ static/js/apps/base/components/header_bar.tsx
