+++ static/js/apps/homepage/components/learn_more.tsx
