+++ static/js/apps/homepage/components/partners.tsx
