+++ static/js/apps/homepage/components/search_animation.tsx
